Move skorch integration to a separate method
