+++ hub/integrations/skorch/utils/__init__.py
@@ -1,1 +1,2 @@
-from .utils import repeat_shape
+from .utils import repeat_image_shape
+from .utils import get_dataset_tensors
